+++ R/validate.R
@@ -16,12 +16,13 @@ validate_csquares <- function(x) {
       x = "'x' is not of class 'csquares'",
       i = "Create a csquares object first with 'new_csquare' or 'as_csquares'"
     ))
-  x <- if(inherits(x, c("character", "vctrs_vctr"))) {
+  x <- if(typeof(x) == "character" || inherits(x, c("character", "vctrs_vctr"))) {
     x
   } else {
     x[[attributes(x)$csquares_col]]
   }
-  strsplit(x, "[|]") |>
+  if (length(x[!is.na(x)]) == 0) return(TRUE)
+  strsplit(x[!is.na(x)], "[|]") |>
     unlist() |>
     .check_csquare_validity() |>
     all()