+++ NEWS.md
@@ -1,3 +1,17 @@
+csquares v0.0.7
+-------------
+
+  * Added experimental functions to convert ICES rectangles
+  * Added tidyverse join operations
+  * Added vignettes:
+    * `vignette("tidy")`
+    * `vignette("csquares-objects")`
+    * `vignette("ices")`
+  * Methods that (potentially) rename columns should update the value for the `csquares_col` attribute.
+    This is now fixed for: `names.csquares()` and `cbind.csquares()`
+  * Implemented `vctrs` cast mechanisms
+  * Better methods for showing and printing `csquares`
+
 csquares v0.0.6
 -------------